pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "local_library"
version = "0.1.0"
description = "A standardized file ingestion tool for Text, Office, and PDF files."
authors = [
  { name="Gherghinoiu Daniel", email="gherghinoiu_daniel@yahoo.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pypdf",
    "python-docx",
    "openpyxl",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = ["tests"]